Save segmentaton image with data type uint8

+++ examples/apps/ai_spleen_seg_app/spleen_seg_operator.py
@@ -27,6 +27,7 @@ from monai.transforms import (
     Spacingd,
     ToTensord,
 )
+from numpy import uint8
 
 
 @input("image", Image, IOType.IN_MEMORY)
@@ -109,6 +110,6 @@ class SpleenSegOperator(Operator):
                 Invertd(
                     keys=pred_key, transform=pre_transforms, orig_keys=self._input_dataset_key, nearest_interp=True
                 ),
-                SaveImaged(keys=pred_key, output_dir=out_dir, output_postfix="seg", resample=False),
+                SaveImaged(keys=pred_key, output_dir=out_dir, output_postfix="seg", output_dtype=uint8, resample=False),
             ]
         )
